+++ fernet_fields/test/settings/pg.py
@@ -1,17 +1,14 @@
 from .base import *  # noqa
 
-import platform
-
-if platform.python_implementation() == 'PyPy':
-    from psycopg2cffi import compat
-    compat.register()
-
 
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql_psycopg2',
-        # matches travis config
+        # matches github actions config
         'NAME': 'djftest',
+        'USER': 'postgres',
+        'PASSWORD': 'postgres',
+        'HOST': 'localhost',
         'TEST': {
             'NAME': 'djftest',
         },